requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==4.2.7
django-crispy-forms==2.1
install==1.3.5
Markdown==3.5.1
Pillow==10.1.0
sqlparse==0.4.4
